companies_api/serialize.py
from rest_framework import serializers

from companies_api.models import Companies


class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Companies
        fields = '__all__'
